more skipToContent refactor
