--- packages/docusaurus-theme-common/src/utils/layoutUtils.tsx
@@ -1,7 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-export const SkipToContentId = 'docusaurus_skipToContent';
